pyproject.toml
[project]
name = "api-ndlocr-lite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dill==0.3.8",
    "fastapi>=0.133.1",
    "flet==0.27.6",
    "lxml==5.4.0",
    "networkx==3.3",
    "numpy==2.2.2",
    "onnxruntime==1.23.2",
    "ordered-set==4.1.0",
    "pillow==12.1.1",
    "protobuf==6.31.1",
    "pyparsing==3.1.2",
    "pypdfium2==4.30.0",
    "python-multipart>=0.0.22",
    "pyyaml==6.0.1",
    "reportlab==4.2.5",
    "tqdm==4.66.4",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
